+++ src/triangulation.h
@@ -44,14 +44,14 @@ public:
 	triangulation(matlab* engine);
 	~triangulation();
 	void create_delaunay(arma::vec& x, arma::vec& y);
-	int get_size();
+	size_t get_size();
 	arma::uvec get_tri(int t);
 private:
 	//triangulation, stored as a list of indexes into the x,y,z data. This is like Matlab
 //	std::vector<std::vector<double> > m_tri;  //size * 3
 
 
-	int m_size;//number of triangulations
+	size_t m_size;//number of triangulations
 
 	std::vector<triangle*> m_triangles;
